Added mTLS - NOT WORKING YET
There is a bug in observatroiu/observatorium which make /-/healthy and
/-/ready endpointes require client-certificate as well as read/write
endpoints therefore k8s fails to deploy the pod.

I have changed `livenessProbe` and `healthnessProbe` to use TCP instead
of HTTP in the mean while and the pod is deployed successfully.

Metrics cannot be delivered to Observatorium because of a
prometheus-operator issue
https://github.com/coreos/prometheus-operator/issues/3118.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>